Error Response sınıfı ile hata mesajları düzeltildi.
