+++ Application/FilmApi.Application/Errors/ErrorResponse.cs
@@ -1,0 +1,28 @@
+﻿namespace FilmApi.Application.Errors;
+
+public class ErrorResponse<T>
+{
+    public bool IsSuccess { get; set; }
+    public string? Message { get; set; }
+    public int StatusCode { get; set; }
+    public T? Data { get; set; }
+
+    public static ErrorResponse<T> Success(T? data = default, string? message = null, int statusCode = 200)
+        => new()
+        {
+            IsSuccess = true,
+            Data = data,
+            Message = message,  
+            StatusCode = statusCode
+        };
+
+    public static ErrorResponse<T> Failure(string message, int statusCode = 400)
+        => new()
+        {
+            IsSuccess = false,
+            Message = message,
+            StatusCode = statusCode
+        };
+
+
+}
